requirements.txt
streamlit 
faiss-cpu 
pandas 
pillow 
tensorflow 
tqdm
matplotlib
opencv-python
numpy